pyproject.toml
[project]
name = "custom-ringtone-generator"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12,<3.13"
dependencies = [
    "musicpy>=7.11",
    "pedalboard>=0.9.17",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "pyfluidsynth>=1.3.4",
    "sf2-loader>=1.25",
]
